Se refactoriza capa service

+++ src/main/java/com/delgadomarset/clinicaOdontologica/service/IOdontologoService.java
@@ -7,9 +7,9 @@ import com.delgadomarset.clinicaOdontologica.entity.Odontologo;
 import java.util.List;
 
 public interface IOdontologoService {
-    OdontologoDto buscarOdontologoPorId(int id);
+    OdontologoDto buscarOdontologoPorId(Long id);
     List<OdontologoDto> listarOdontologos();
     OdontologoDto registrarOdontologo(Odontologo odontologo);
     OdontologoDto actualizarOdontologo(Odontologo odontologo);
-    void eliminarOdontologo(int id);
+    void eliminarOdontologo(Long id);
 }
